pyproject.toml
[project]
name = "python-grpc-server-langchain-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.11.*"
dependencies = [
    "grpcio",
    "grpcio-tools==1.62.0",
    "protobuf==4.25.3",
    "grpcio-health-checking",
    "langchain",
    "langchain-google-genai",
    "python-dotenv",
    "google-generativeai",
]
